feat(scatterplot): move stories to the dedicated storybook workspace
