+++ storybook/package.json
@@ -32,6 +32,7 @@
     "@nivo/radar": "workspace:*",
     "@nivo/radial-bar": "workspace:*",
     "@nivo/sankey": "workspace:*",
+    "@nivo/scatterplot": "workspace:*",
     "@nivo/tooltip": "workspace:*",
     "@nivo/waffle": "workspace:*",
     "@babel/preset-env": "^7.21.5",
@@ -51,6 +52,8 @@
     "storybook": "^7.0.7",
     "d3-scale": "^3.2.3",
     "@types/d3-scale": "^3.2.3",
+    "d3-shape": "^1.3.5",
+    "@types/d3-shape": "^2.0.0",
     "styled-components": "^5.3.3",
     "d3-random": "^1.1.2",
     "@types/d3-random": "^1.1.3",
